PR4: Move ConstitutiveModels/ to Materials/Constitutive/
- Move all 11 constitutive model files to Materials/Constitutive directory
- Update imports in __init__.py and Material.py to reference new location
- Add legacy path alias in extensions.py to enable extension overlay compatibility
- Preserve PascalCase filenames and relative imports (no auto-discovery changes needed)
