+++ safeincave/Materials/Material.py
@@ -5,10 +5,10 @@
 from __future__ import annotations
 import torch as to
 
-# Import all constitutive models from the ConstitutiveModels folder
-from ..ConstitutiveModels import Spring, Thermoelastic
-from ..ConstitutiveModels import *  # noqa: F403, F405
-from ..ConstitutiveModels import __all__ as _CONSTITUTIVE_ALL
+# Import all constitutive models from the Constitutive subfolder
+from .Constitutive import Spring, Thermoelastic
+from .Constitutive import *  # noqa: F403, F405
+from .Constitutive import __all__ as _CONSTITUTIVE_ALL
 
 __all__ = ["Material"] + _CONSTITUTIVE_ALL
